pyproject.toml
[project]
name = "photo-pipeline"
version = "0.1.0"
description = "Privacy-first photo sovereignty pipeline with EXIF/GPS extraction, object detection, and semantic search"
authors = [
    {name = "melbourne-quantum-dev", email = "leonardo97@tuta.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0.0",
    "pillow-heif>=0.13.0",
    "pyyaml>=6.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ipython>=8.0.0",
    "ruff>=0.1.0",
]
stage3 = [
    "ultralytics>=8.0.0",  # YOLO object detection
    "torch>=2.0.0",
]
stage4 = [
    "open-clip-torch>=2.20.0",  # CLIP semantic embeddings
]
stage5 = [
    "easyocr>=1.7.0",  # OCR text extraction
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
# Exclude archived test scripts from discovery
norecursedirs = ["tests/archive", ".git", ".venv", "__pycache__"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py
